Ground-up rewrite!
A complete rewrite with inspiration and code from broccoli's own
CLI module. Broccoli ^1.0.0 is now used, as well as Grunt ^1.0.0 for
tests.

See CHANGELOG.md for a more exhaustive list of changes. The API is
mostly compatible with previous versions, with the exception that
task options have been moved into the options object.